Fixed errors , added typescript
