+++ README.md
@@ -5,3 +5,16 @@ An e-commerce webapp built using react
 [visit to store now!!!](https://fauxica-sohan9819.vercel.app)
 
 ![preview](https://github.com/sohan9819/fauxica/assets/64985447/080ee6fe-608c-4206-a256-97d3b574b29c)
+
+# Test Card for Indian Payments
+
+| Card Network | Card Number         | CVV        | Expiry Date     |
+| ------------ | ------------------- | ---------- | --------------- |
+| Mastercard   | 5267 3181 8797 5449 | Random CVV | Any future date |
+| Visa         | 4111 1111 1111 1111 | Random CVV | Any future date |
+
+# Test Card for EMI Payments
+
+| Card Network | Card Number         | CVV        | Expiry Date     |
+| ------------ | ------------------- | ---------- | --------------- |
+| Mastercard   | 5241 8100 0000 0000 | Random CVV | Any future date |
